requirements.txt
Flask==3.0.0
requests==2.31.0
azure-identity==1.15.0
python-dotenv==1.0.0